+++ lib/rh-admin-utils/PageRestrictions.php
@@ -201,7 +201,7 @@ class PageRestrictions
     public static function before_delete_post(int $post_id, \WP_Post $post): void
     {
         if (self::is_locked($post)) {
-            wp_die(__("Can't delete post <strong>#$post->ID</strong> as it is locked."));
+            wp_die(__("Can't delete post <strong>{$post->post_title} (#{$post->ID})</strong> as it is locked."));
         }
     }